reorganize files, fix issue with parsing arc segments into commands, make util functions take commands array for reusability
