+++ src/common/PathData/lib/geom.ts
