--- src/common/PathData/module.ts
@@ -1,5 +1,0 @@
-export { PathData } from './index';
-export * from './geom';
-export * from './helpers';
-export * from './matrix';
-export * from './types';
