+++ src/common/util/svg.ts
@@ -1,7 +1,7 @@
 import React from 'react';
 import { svgPathBbox } from 'svg-path-bbox';
 import { Dimensions } from '@/common/util/data';
-import { PathData } from '@/common/PathData/module';
+import { PathData } from '@/common/PathData';
 
 const parseString = (str) => {
   const parser = new window.DOMParser();
