+++ src/widgets/CrosshatchShelves/TriangularGrid/index.tsx
@@ -6,7 +6,7 @@ import { ExtendedModel } from 'mobx-keystone';
 import { LicenseWatermarkContent } from '@/widgets/LicenseWatermarkContent';
 import { BaseWidgetClass } from '@/WidgetWorkspace/widget-types/BaseWidgetClass';
 import { assertNotNullish } from '@/common/util/assert';
-import { PathData } from '@/common/PathData/module';
+import { PathData } from '@/common/PathData';
 import { numberTextProp, sliderWithTextProp, switchProp } from '../../../common/keystone-tweakables/props';
 import { PIXELS_PER_INCH } from '../../../common/util/units';
 import {
