+++ src/widgets/CrosshatchShelves/util.ts
@@ -1,6 +1,6 @@
 import Flatten from '@flatten-js/core';
 import { assertNotNullish } from '@/common/util/assert';
-import { PathData } from '@/common/PathData/module';
+import { PathData } from '@/common/PathData';
 import { hingedPlot, polygonWithFace } from '../../common/util/geom';
 import Segment = Flatten.Segment;
 import Point = Flatten.Point;
