+++ src/widgets/PyramidNet/components/TextureEditorDrawer/components/TextureEditor/models/TextureEditorModel.tsx
@@ -6,9 +6,9 @@ import {
 import {
   action, computed, makeObservable, observable,
 } from 'mobx';
-import { RawPoint } from '@/common/PathData/module';
 import { Dimensions } from '@/common/util/data';
 import { assertNotNullish } from '@/common/util/assert';
+import { RawPoint } from '@/common/PathData';
 import { BoundaryModel } from './BoundaryModel';
 import { ModifierTrackingModel } from './ModifierTrackingModel';
 import {
